requirements.txt
asgiref==3.10.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.9.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
gunicorn==22.0.0
packaging==25.0
psycopg2-binary==2.9.7
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
